covert to float64

+++ metrics_system.go
@@ -36,22 +36,22 @@ func (e *HarborExporter) collectSystemMetric(ch chan<- prometheus.Metric) bool {
 
 	// Set all bool values as separate metrics
 	ch <- prometheus.MustNewConstMetric(
-		allMetrics["system_with_notary"].Desc, allMetrics["system_with_notary"].Type, Btoi(data.WithNotary),
+		allMetrics["system_with_notary"].Desc, allMetrics["system_with_notary"].Type, float64(Btoi(data.WithNotary)),
 	)
 	ch <- prometheus.MustNewConstMetric(
-		allMetrics["system_self_registration"].Desc, allMetrics["system_self_registration"].Type, Btoi(data.SelfRegistration),
+		allMetrics["system_self_registration"].Desc, allMetrics["system_self_registration"].Type, float64(Btoi(data.SelfRegistration)),
 	)
 	ch <- prometheus.MustNewConstMetric(
-		allMetrics["system_has_ca_root"].Desc, allMetrics["system_has_ca_root"].Type, Btoi(data.HasCaRoot),
+		allMetrics["system_has_ca_root"].Desc, allMetrics["system_has_ca_root"].Type, float64(Btoi(data.HasCaRoot)),
 	)
 	ch <- prometheus.MustNewConstMetric(
-		allMetrics["system_read_only"].Desc, allMetrics["system_read_only"].Type, Btoi(data.ReadOnly),
+		allMetrics["system_read_only"].Desc, allMetrics["system_read_only"].Type, float64(Btoi(data.ReadOnly)),
 	)
 	ch <- prometheus.MustNewConstMetric(
-		allMetrics["system_with_chartmuseum"].Desc, allMetrics["system_with_chartmuseum"].Type, Btoi(data.WithChartmuseum),
+		allMetrics["system_with_chartmuseum"].Desc, allMetrics["system_with_chartmuseum"].Type, float64(Btoi(data.WithChartmuseum)),
 	)
 	ch <- prometheus.MustNewConstMetric(
-		allMetrics["system_notification_enable"].Desc, allMetrics["system_notification_enable"].Type, Btoi(data.NotificationEnable),
+		allMetrics["system_notification_enable"].Desc, allMetrics["system_notification_enable"].Type, float64(Btoi(data.NotificationEnable)),
 	)
 
 	return true
